+++ Example/Example controllers/ReorderViewController.swift
@@ -31,7 +31,7 @@ class ReorderViewController: UIViewController, DTTableViewManageable, UITableVie
         manager.editingStyle(for: StringCell.self, { _,_,_ in return .none })
         
         manager.memoryStorage.setSectionHeaderModels(["Section 1", "Section 2", "Section 3"])
-        manager.move(StringCell.self) { [weak manager] _, _, source, destination in
+        manager.move(StringCell.self) { [weak manager] destination, _, _, source in
             manager?.memoryStorage.moveItemWithoutAnimation(from: source, to: destination)
         }