+++ Source/DTTableViewManager+Drag.swift
@@ -41,7 +41,7 @@ extension DTTableViewManager
     }
     
     @available(iOS 11, *)
-    // Registers `closure` to be executed when `UITableViewDragDelegate.tableView(_:itemsForAddingTo:at:point:)` method is called for `cellClass`
+    /// Registers `closure` to be executed when `UITableViewDragDelegate.tableView(_:itemsForAddingTo:at:point:)` method is called for `cellClass`
     open func itemsForAddingToDragSession<T:ModelTransfer>(from cellClass: T.Type, _ closure: @escaping (UIDragSession, CGPoint, T, T.ModelType, IndexPath) -> [UIDragItem]) where T: UITableViewCell
     {
         tableDragDelegate?.append5ArgumentReaction(for: T.self,